[#1745] Move owning pointer to elementary; rename other pointer types
