+++ iceoryx2-bb/container/src/flatmap.rs
@@ -40,8 +40,8 @@ use core::mem::MaybeUninit;
 use iceoryx2_bb_concurrency::atomic::AtomicBool;
 use iceoryx2_bb_elementary::CallbackProgression;
 use iceoryx2_bb_elementary::bump_allocator::BumpAllocator;
-use iceoryx2_bb_elementary::relocatable_ptr::GenericRelocatablePointer;
-use iceoryx2_bb_elementary_traits::owning_pointer::GenericOwningPointer;
+use iceoryx2_bb_elementary::owning_pointer::GenericOwningPointer;
+use iceoryx2_bb_elementary::relocatable_pointer::GenericRelocatablePointer;
 use iceoryx2_bb_elementary_traits::pointer_family::{NonNullFamily, PointerFamily};
 pub use iceoryx2_bb_elementary_traits::relocatable_container::RelocatableContainer;
 use iceoryx2_bb_elementary_traits::{
