+++ iceoryx2-bb/container/src/queue.rs
@@ -122,9 +122,9 @@ use core::{alloc::Layout, fmt::Debug, mem::MaybeUninit};
 use iceoryx2_bb_concurrency::atomic::AtomicBool;
 use iceoryx2_bb_elementary::bump_allocator::BumpAllocator;
 use iceoryx2_bb_elementary::math::unaligned_mem_size;
-use iceoryx2_bb_elementary::relocatable_ptr::{GenericRelocatablePointer, RelocatablePointer};
+use iceoryx2_bb_elementary::owning_pointer::{GenericOwningPointer, OwningPointer};
+use iceoryx2_bb_elementary::relocatable_pointer::{GenericRelocatablePointer, RelocatablePointer};
 use iceoryx2_bb_elementary_traits::allocator::{AllocationError, BaseAllocator};
-use iceoryx2_bb_elementary_traits::owning_pointer::{GenericOwningPointer, OwningPointer};
 use iceoryx2_bb_elementary_traits::placement_default::PlacementDefault;
 use iceoryx2_bb_elementary_traits::pointer::Pointer;
 use iceoryx2_bb_elementary_traits::pointer_family::{NonNullFamily, PointerFamily};
