+++ iceoryx2-bb/container/src/slotmap.rs
@@ -53,8 +53,8 @@ use core::mem::MaybeUninit;
 use iceoryx2_bb_concurrency::atomic::AtomicBool;
 use iceoryx2_bb_derive_macros::ZeroCopySend;
 use iceoryx2_bb_elementary::bump_allocator::BumpAllocator;
-use iceoryx2_bb_elementary::relocatable_ptr::GenericRelocatablePointer;
-use iceoryx2_bb_elementary_traits::owning_pointer::GenericOwningPointer;
+use iceoryx2_bb_elementary::owning_pointer::GenericOwningPointer;
+use iceoryx2_bb_elementary::relocatable_pointer::GenericRelocatablePointer;
 use iceoryx2_bb_elementary_traits::placement_default::PlacementDefault;
 use iceoryx2_bb_elementary_traits::pointer_family::{NonNullFamily, PointerFamily};
 pub use iceoryx2_bb_elementary_traits::relocatable_container::RelocatableContainer;
