+++ iceoryx2-bb/container/src/vector/relocatable_vec.rs
@@ -78,7 +78,7 @@ use core::{
 };
 use core::{fmt::Debug, mem::MaybeUninit};
 
-use iceoryx2_bb_elementary::{math::unaligned_mem_size, relocatable_ptr::*};
+use iceoryx2_bb_elementary::{math::unaligned_mem_size, relocatable_pointer::*};
 pub use iceoryx2_bb_elementary_traits::relocatable_container::RelocatableContainer;
 use iceoryx2_bb_elementary_traits::{pointer_family::NonNullFamily, zero_copy_send::ZeroCopySend};
 use iceoryx2_log::{fail, fatal_panic};
