+++ iceoryx2-bb/elementary-traits/src/lib.rs
@@ -22,7 +22,6 @@ mod as_cstr;
 pub use crate::as_cstr::*;
 pub mod atomic_copy;
 pub mod iceoryx_send;
-pub mod owning_pointer;
 pub mod placement_default;
 pub mod plain_old_data_without_padding;
 pub mod pointer;
