+++ iceoryx2-bb/elementary/tests-common/src/relocatable_ptr_tests.rs
@@ -10,7 +10,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use iceoryx2_bb_elementary::relocatable_ptr::RelocatablePointer;
+use iceoryx2_bb_elementary::relocatable_pointer::RelocatablePointer;
 use iceoryx2_bb_elementary_traits::pointer::Pointer;
 use iceoryx2_bb_testing::assert_that;
 use iceoryx2_bb_testing_macros::test;
