+++ iceoryx2-bb/lock-free/src/mpmc/bit_set.rs
@@ -42,14 +42,13 @@ use core::{alloc::Layout, fmt::Debug};
 
 use iceoryx2_bb_concurrency::atomic::Ordering;
 use iceoryx2_bb_concurrency::atomic::{AtomicBool, AtomicU8, AtomicUsize};
+use iceoryx2_bb_elementary::owning_pointer::OwningPointer;
 use iceoryx2_bb_elementary::{
     bump_allocator::BumpAllocator,
     math::unaligned_mem_size,
-    relocatable_ptr::{Pointer, RelocatablePointer},
+    relocatable_pointer::{Pointer, RelocatablePointer},
 };
-use iceoryx2_bb_elementary_traits::{
-    owning_pointer::OwningPointer, relocatable_container::RelocatableContainer,
-};
+use iceoryx2_bb_elementary_traits::relocatable_container::RelocatableContainer;
 
 use iceoryx2_log::{fail, fatal_panic};
